backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
httpx==0.28.1
python-dotenv==1.1.0
pydantic==2.11.5
pydantic-settings==2.9.1
modal==1.0.3
boto3==1.38.36
supabase==2.15.3
google-auth==2.25.2
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
PyJWT==2.8.0
